+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add clippy warning for complex functions.
 - Code refactoring to make it easier to understand.
+- Extract library part into crate gleisbau
 
 ### Removed